Added service loginVerified which contains loginModel related rpcs which require authentication
