+++ go.mod
@@ -5,7 +5,7 @@ go 1.19
 require (
 	github.com/SaiNageswarS/go-api-boot v0.1.36-0.20240302074729-626c6b5bab97
 	github.com/google/uuid v1.5.0
-	github.com/grpc-ecosystem/go-grpc-middleware v1.3.0
+	github.com/grpc-ecosystem/go-grpc-middleware v1.4.0
 	github.com/jinzhu/copier v0.3.2
 	github.com/rs/cors v1.9.0
 	github.com/twilio/twilio-go v0.9.0
