+++ service/profile_service.go
@@ -47,11 +47,6 @@ func (s *ProfileService) CreateOrUpdateProfile(ctx context.Context, req *pb.Crea
 	// get existing profile
 	oldProfile := getExistingOrEmptyProfile(s.db, tenant, userId)
 
-	// check if profile is marked for deletion
-	if oldProfile != nil && oldProfile.DeletionInfo.MarkedForDeletion {
-		return nil, status.Error(codes.PermissionDenied, "Profile marked for deletion")
-	}
-
 	isNewUser := false
 	if len(oldProfile.UserId) == 0 {
 		isNewUser = true
@@ -169,199 +164,6 @@ func (s *ProfileService) BulkGetProfileByIds(ctx context.Context, req *pb.BulkGe
 		logger.Error("Failed getting profile", zap.Error(err))
 		return nil, status.Error(codes.Internal, "Failed getting profiles")
 	}
-}
-
-func (s *ProfileService) RequestProfileDeletion(ctx context.Context, req *pb.ProfileDeletionRequest) (*pb.StatusResponse, error) {
-	userId, tenant := auth.GetUserIdAndTenant(ctx)
-
-	// Fetch profile info
-	profileResChan, errChan := s.db.Profile(tenant).FindOneById(userId)
-
-	select {
-	case profileRes := <-profileResChan:
-		// Save profile deletion request
-		profileRes.DeletionInfo = models.DeletionInfo{
-			MarkedForDeletion: true,
-			DeletionTime:      time.Now().Unix(),
-			Reason:            req.Reason,
-		}
-		err := <-s.db.Profile(tenant).Save(profileRes)
-
-		if err != nil {
-			logger.Error("Failed saving profile deletion request", zap.Error(err))
-			return nil, status.Error(codes.Internal, "Failed saving profile deletion request")
-		}
-	case err := <-errChan:
-		logger.Error("Failed getting profile", zap.Error(err))
-		return nil, status.Error(codes.Internal, "Failed getting profile")
-	}
-
-	return &pb.StatusResponse{
-		Status: "Profile deletion request sent successfully",
-	}, nil
-
-}
-
-// GetPendingProfileDeletionRequests returns all profiles marked for deletion and is used by admin only.
-func (s *ProfileService) GetPendingProfileDeletionRequests(ctx context.Context, req *pb.GetProfileDeletionRequest) (*pb.ProfileListResponse, error) {
-	userID, tenant := auth.GetUserIdAndTenant(ctx)
-
-	// Check if user is admin
-	if !s.db.Login(tenant).IsAdmin(userID) {
-		return nil, status.Error(codes.PermissionDenied, "User with id "+userID+" don't have permission")
-	}
-
-	// Fetch pending profile deletion requests
-	filter := bson.M{
-		"deletionInfo.markedForDeletion": true,
-	}
-
-	if req.PageSize == 0 {
-		req.PageSize = 10
-	}
-
-	skip := int64(req.PageNumber * req.PageSize)
-	profileDeletionRequestsChan, errChan := s.db.Profile(tenant).Find(filter, nil, int64(req.PageSize), skip)
-	totalCountResChan, countErrChan := s.db.Profile(tenant).CountDocuments(filter)
-
-	// get total count of pending profile deletion requests
-	totalCount := 0
-	select {
-	case count := <-totalCountResChan:
-		totalCount = int(count)
-	case err := <-countErrChan:
-		logger.Error("Error fetching user count", zap.Error(err))
-	}
-
-	// convert profile model to proto
-	var profileProtoList []*pb.UserProfileProto
-	var userIds []string
-	select {
-	case profileDeletionRequests := <-profileDeletionRequestsChan:
-		for _, profile := range profileDeletionRequests {
-			profileProtoList = append(profileProtoList, getProfileProto(&profile))
-			userIds = append(userIds, profile.UserId)
-		}
-
-	case err := <-errChan:
-		logger.Error("Failed getting profile deletion requests", zap.Error(err))
-		return nil, status.Error(codes.Internal, "Failed getting profile deletion requests")
-	}
-
-	// get login info using userId
-	loginInfoChan, errChan := s.db.Login(tenant).FindByIds(userIds)
-
-	// populate phone number field in profile proto
-	var loginInfo []models.LoginModel
-	select {
-	case loginInfo = <-loginInfoChan:
-	case <-errChan:
-		logger.Error("Failed getting login info")
-	}
-
-	if len(loginInfo) > 0 {
-		populateLoginInfo(profileProtoList, loginInfo)
-	}
-
-	return &pb.ProfileListResponse{
-		Profiles:   profileProtoList,
-		TotalUsers: int64(totalCount),
-	}, nil
-}
-
-// DeleteProfile deletes profile and login from db and is used by admin only.
-func (s *ProfileService) DeleteProfile(ctx context.Context, req *pb.IdRequest) (*pb.StatusResponse, error) {
-	userId, tenant := auth.GetUserIdAndTenant(ctx)
-
-	// Check if user is admin
-	if !s.db.Login(tenant).IsAdmin(userId) {
-		return nil, status.Error(codes.PermissionDenied, "User with id "+userId+" don't have permission")
-	}
-
-	// Check if profile exists
-	isExists := s.db.Profile(tenant).IsExistsById(req.UserId)
-
-	if !isExists {
-		return &pb.StatusResponse{
-			Status: "Profile not found",
-		}, nil
-	}
-
-	// Delete profile from db
-	err := <-s.db.Profile(tenant).DeleteById(req.UserId)
-	if err != nil {
-		logger.Error("Failed deleting profile", zap.Error(err))
-		return nil, status.Error(codes.Internal, "Failed deleting profile")
-	}
-
-	// Delete login from db
-	err = <-s.db.Login(tenant).DeleteById(req.UserId)
-	if err != nil {
-		logger.Error("Failed deleting login", zap.Error(err))
-		return nil, status.Error(codes.Internal, "Failed deleting login")
-	}
-
-	// TODO: Delete all posts, comments, notifications, etc. related to this user.
-
-	return &pb.StatusResponse{
-		Status: "Profile deleted successfully",
-	}, nil
-}
-
-// CancelProfileDeletionRequest cancels profile deletion request and is used by admin only.
-func (s *ProfileService) CancelProfileDeletionRequest(ctx context.Context, req *pb.IdRequest) (*pb.StatusResponse, error) {
-	userId, tenant := auth.GetUserIdAndTenant(ctx)
-
-	// Check if user is admin
-	if !s.db.Login(tenant).IsAdmin(userId) {
-		return nil, status.Error(codes.PermissionDenied, "User with id "+userId+" don't have permission")
-	}
-
-	// Fetch profile info
-	profileResChan, errChan := s.db.Profile(tenant).FindOneById(req.UserId)
-
-	select {
-	case profileRes := <-profileResChan:
-		// Un mark profile for deletion
-		profileRes.DeletionInfo = models.DeletionInfo{
-			MarkedForDeletion: false,
-			DeletionTime:      0,
-			Reason:            "",
-		}
-		err := <-s.db.Profile(tenant).Save(profileRes)
-
-		if err != nil {
-			logger.Error("Failed saving profile deletion request", zap.Error(err))
-			return nil, status.Error(codes.Internal, "Failed saving profile deletion request")
-		}
-	case err := <-errChan:
-		logger.Error("Failed getting profile", zap.Error(err))
-		return nil, status.Error(codes.Internal, "Failed getting profile")
-	}
-
-	return &pb.StatusResponse{
-		Status: "Profile deletion request cancelled successfully",
-	}, nil
-}
-
-// check if user is admin or not and return response. Used by admin only.
-func (s *ProfileService) IsUserAdmin(ctx context.Context, req *pb.IdRequest) (*pb.IsUserAdminResponse, error) {
-	userId, tenant := auth.GetUserIdAndTenant(ctx)
-
-	if len(req.UserId) > 0 {
-		userId = req.UserId
-	}
-
-	// Check if user is admin
-	if !s.db.Login(tenant).IsAdmin(userId) {
-		return nil, status.Error(codes.PermissionDenied, "User with id "+userId+" don't have permission")
-	}
-
-	isAdmin := s.db.Login(tenant).IsAdmin(userId)
-
-	return &pb.IsUserAdminResponse{
-		IsAdmin: isAdmin,
-	}, nil
 }
 
 func (s *ProfileService) GetProfileImageUploadUrl(ctx context.Context, req *pb.ProfileImageUploadRequest) (*pb.ProfileImageUploadURL, error) {
@@ -512,68 +314,6 @@ func (s *ProfileService) ChangeUserType(ctx context.Context, req *pb.ChangeUserT
 	}, nil
 }
 
-func (s *ProfileService) BlockUser(ctx context.Context, req *pb.IdRequest) (*pb.StatusResponse, error) {
-	userId, tenant := auth.GetUserIdAndTenant(ctx)
-
-	// Check if user is admin
-	if !s.db.Login(tenant).IsAdmin(userId) {
-		return nil, status.Error(codes.PermissionDenied, "User with id "+userId+" don't have permission")
-	}
-
-	// fetch profile info
-	profileResChan, errChan := s.db.Profile(tenant).FindOneById(req.UserId)
-
-	select {
-	case profileRes := <-profileResChan:
-		// block user
-		profileRes.IsBlocked = true
-		err := <-s.db.Profile(tenant).Save(profileRes)
-
-		if err != nil {
-			logger.Error("Failed blocking user", zap.Error(err))
-			return nil, status.Error(codes.Internal, "Failed blocking user")
-		}
-	case err := <-errChan:
-		logger.Error("Failed getting profile", zap.Error(err))
-		return nil, status.Error(codes.Internal, "Failed getting profile")
-	}
-
-	return &pb.StatusResponse{
-		Status: "User blocked successfully",
-	}, nil
-}
-
-func (s *ProfileService) UnblockUser(ctx context.Context, req *pb.IdRequest) (*pb.StatusResponse, error) {
-	userId, tenant := auth.GetUserIdAndTenant(ctx)
-
-	// Check if user is admin
-	if !s.db.Login(tenant).IsAdmin(userId) {
-		return nil, status.Error(codes.PermissionDenied, "User with id "+userId+" don't have permission")
-	}
-
-	// fetch profile info
-	profileResChan, errChan := s.db.Profile(tenant).FindOneById(req.UserId)
-
-	select {
-	case profileRes := <-profileResChan:
-		// unblock user
-		profileRes.IsBlocked = false
-		err := <-s.db.Profile(tenant).Save(profileRes)
-
-		if err != nil {
-			logger.Error("Failed unblocking user", zap.Error(err))
-			return nil, status.Error(codes.Internal, "Failed unblocking user")
-		}
-	case err := <-errChan:
-		logger.Error("Failed getting profile", zap.Error(err))
-		return nil, status.Error(codes.Internal, "Failed getting profile")
-	}
-
-	return &pb.StatusResponse{
-		Status: "User unblocked successfully",
-	}, nil
-}
-
 // get profile proto from profile model
 func getProfileProto(profileModel *models.ProfileModel) *pb.UserProfileProto {
 	result := &pb.UserProfileProto{}
@@ -664,16 +404,4 @@ func getExistingOrEmptyProfile(db db.AuthDbInterface, tenant, userId string) *mo
 	}
 
 	return profile
-}
-
-func populateLoginInfo(userProfileProto []*pb.UserProfileProto, loginInfo []models.LoginModel) {
-	for i, profile := range userProfileProto {
-		for _, loginModel := range loginInfo {
-			if profile.UserId == loginModel.UserId {
-				userProfileProto[i].PhoneNumber = loginModel.Phone
-				copier.Copy(&userProfileProto[i], &loginModel)
-				break
-			}
-		}
-	}
 }
